Extract page comments panel
